JYCAPP_EN-US/pyproject.toml
[tool.poetry]
name = "jyc"
version = "0.1.3"
description = "A configuration file handling library supporting INI, JSON, and YAML formats"
authors = ["yubo <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^6.0"
configparser = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "./*"

[tool.poetry.scripts]
jyc = "app:main"
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

